+++ src/main/java/org/dinosaur/foodbowl/domain/store/application/dto/StoreCreateDto.java
@@ -1,0 +1,17 @@
+package org.dinosaur.foodbowl.domain.store.application.dto;
+
+import java.math.BigDecimal;
+
+public record StoreCreateDto(
+        String storeName,
+        String category,
+        String address,
+        BigDecimal storeX,
+        BigDecimal storeY,
+        String storeUrl,
+        String phone,
+        String schoolName,
+        BigDecimal schoolX,
+        BigDecimal schoolY
+) {
+}
